QuaC Pulse Visitor now supports digital gates to pulses translation
This is a two-stage process: visit all the instructions, if not pulses, then generate the corresponding pulse cmd-defs.

This translation is very simple at this stage: check if the cmd-def is explicitly defined, otherwise, use the universal U3 cmd-def. Hence, this requires, at the very mimimum, the pulse library must contain the CNOT and U3 pulse cmd-defs.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>