Have the option to disable PETSc adaptive time-stepping
This is the root-cause of inconsistent results when we simulating pulses (mixed with LO signals)
Somehow, the adaptive time-stepping just doesn't cope well with this type of simulation, perhaps, it prefers the *usual* set-up where we are in the rotating frames (it passes all the qutip-like tests where we have this type of set-up).

By fixing this, we can now introduce the real *Pulse* test suite.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>