Add a universal U3 fallback to digital simulation
The QuaC visitor is structured in such a way that digital gate will be transformed to pulses via a 2-tier look-up:

- First, it looks for pulse cmd-def for that particular gate (e.g. pulse::h_0 for Hadamard on qubit 0).

- If not, it will try the u3 cmd_def which is equivalent to that gate.

We  add another fallback: if the *pulse* cmd for U3 does not exist, use a *digital* U3, i.e. direct linear algebra simulation of that U3 gate in QuaC.

This is useful for us now since we are still building up our pulse library.
One extra bonus is that it enables us to test the gate (H, Rz, Rx, X, Y, Z, etc.) to U3 conversion procedure (implemented but never tested until now)

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>