Initial work on Python bindings
The System Model is a bit tricky for binding because the Python wrapper for heterogeneous map has a set of known variant types and System Model doesn't exist in XACC.

Solution for this is to use the service registry as a mechanism to communicate the system model: i.e. on the Python side, we get a reference to the system model via the service registry and initialize the configs. The QuaC accelerator then has a mechanism to retrieve the configs from the service registry.

This is still WIP, will test and provide examples.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>