Fix QuaC solver: product terms of the same qubit subspace
Problem: Hamiltonian terms like O1*O1 (product of number operators on the same qubit) were not handled correctly by QuaC
I think it was not designed to handle that.

This explained weird results that I got when using the transmon qubit Hamiltonian in that form.

Fix: I added a completely independent code path to handle this case, starting at the XACC-QuaC interface API so that this is an isolated fix.

Added the unit test checks now that the results make sense.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>